requirements.txt
flask
openai
dotenv
Pillow